+++ internal/controller/replicationdestination_test.go
@@ -32,6 +32,10 @@ var _ = Describe("ReplicationDestination", func() {
 		namespace = &corev1.Namespace{
 			ObjectMeta: metav1.ObjectMeta{
 				GenerateName: "volsync-test-",
+				Annotations: map[string]string{
+					// Rsync-ssh requires privileged movers enabled via this ns annotation
+					volsyncv1alpha1.PrivilegedMoversNamespaceAnnotation: "true",
+				},
 			},
 		}
 		createWithCacheReload(ctx, k8sClient, namespace)
@@ -80,6 +84,46 @@ var _ = Describe("ReplicationDestination", func() {
 			Expect(errCond.Status).To(Equal(metav1.ConditionFalse))
 			Expect(errCond.Reason).To(Equal(volsyncv1alpha1.SynchronizingReasonError))
 			Expect(errCond.Message).To(ContainSubstring("a replication method must be specified"))
+		})
+	})
+
+	//nolint:dupl
+	Context("When the privileged movers annotation is not set on the ns", func() {
+		var unprivilegedNamespace *corev1.Namespace
+
+		BeforeEach(func() {
+			// create a separate namespace without the privileged movers annotation for this test
+			// Each test is run in its own namespace
+			unprivilegedNamespace = &corev1.Namespace{
+				ObjectMeta: metav1.ObjectMeta{
+					GenerateName: "volsync-test-unprivileged-",
+				},
+			}
+			createWithCacheReload(ctx, k8sClient, unprivilegedNamespace)
+			Expect(unprivilegedNamespace.Name).NotTo(BeEmpty())
+
+			// Update our RD to use the unprivileged namespace
+			rd.Namespace = unprivilegedNamespace.Name
+
+			// Set rsync as the replication method
+			rd.Spec.Rsync = &volsyncv1alpha1.ReplicationDestinationRsyncSpec{}
+		})
+		AfterEach(func() {
+			// All resources are namespaced, so this should clean it all up
+			Expect(k8sClient.Delete(ctx, unprivilegedNamespace)).To(Succeed())
+		})
+
+		It("the CR should report an error in the status (requiring privileged mode)", func() {
+			Eventually(func() *volsyncv1alpha1.ReplicationDestinationStatus {
+				Expect(k8sClient.Get(ctx, client.ObjectKeyFromObject(rd), rd)).To(Succeed())
+				return rd.Status
+			}, duration, interval).ShouldNot(BeNil())
+			Expect(rd.Status.Conditions).To(HaveLen(1))
+			errCond := rd.Status.Conditions[0]
+			Expect(errCond.Type).To(Equal(volsyncv1alpha1.ConditionSynchronizing))
+			Expect(errCond.Status).To(Equal(metav1.ConditionFalse))
+			Expect(errCond.Reason).To(Equal(volsyncv1alpha1.SynchronizingReasonError))
+			Expect(errCond.Message).To(ContainSubstring("mover requires privileged mode"))
 		})
 	})